+++ src/help.rs
@@ -1,5 +1,5 @@
 /// Help text.
-pub const HELP: &str = "# felix v1.0.1
+pub const HELP: &str = "# felix v1.1.0
 A simple TUI file manager with vim-like keymapping.
 
 ## Usage
@@ -44,6 +44,11 @@ Esc               :Return to the normal mode.
 :h                :Show help.
 :q / ZZ           :Exit.
 
+## Preview feature
+By default, text files and directories can be previewed.
+To preview images, you need to install chafa.
+Please see https://hpjansson.org/chafa/
+
 ## Configuration
 ### Linux
 config file    : $XDG_CONFIG_HOME/felix/config.toml
@@ -55,5 +60,5 @@ config file    : $HOME/Library/Application Support/felix/config.toml
 trash directory: $HOME/Library/Application Support/felix/trash
 log files      : $HOME/Library/Application Support/felix/log
 
-For more detail, visit https://github.com/kyoheiu/felix
+For more details, visit https://github.com/kyoheiu/felix
 ";