+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
-## [Unreleased]
-
 ## Notes
 
-- Trying to make a user guide (just to show and explain how to use each commands) by `mdbook`.
 - Text preview color needs to be readable enough, so it's worth rethinking (Now LightBlack).
+
+## [Unreleased]
+
+## v1.1.0 (2022-08-08)
+
+### Changed
+
+- Use `chafa` instead of `libsixel` & `viuer` to preview image files. This greatly improves the performance and code maintainability, and as a consequence, the number of dependencies is reduced (137 -> 53).
+- With `chafa`, the hi-res image preview is supported in kitty or terminals that support sixel.
+- Files larger than 1GB are no longer previewed in order to improve the performance.
+- Remove profile.release to support older version of Rust.
+- Huge refactoring (layout.rs created).
+
+### Added
+
+- `content-inspector` to exclude binary files to be previewed.
 
 ## v1.0.1 (2022-07-28)